Fix (un)folding in 2.8x
It worked previously only if there was only one mesh rigged to the targeted armature.
With (at least) two meshes, I guess that when processing the second mesh and updating the depsgraph it somehow invalidated the previous to_mesh() result, even though that was a copy()
(and then Blender crashed with an access violation error)
Discovered bpy.data.meshes.new_from_object() by accident, replaced to_mesh with it, works like a charm!