Fix potential 2.8x issue about modifiers not being applied
The disable-modifier logic is in charge of syncing the show-viewport and show-render properties so that in 2.8x with the removal of the to_mesh settings about using viewport/render settings for modifiers, it was certain that a modifier would be applied or not according to export options.
But it was also as a whole under a condition of the mesh being rigged because that's what the disable-modifier logic was for initially.
The fix moves the disable-armature-modifier condition into a more appropriate place.